+++ src/graphology/tests/read.js
@@ -296,6 +296,247 @@ export default function read(Graph, checkers) {
       }
     },
 
+    '#.areDirectedNeighbors': {
+      'it should throw if node is not in the graph.': function () {
+        const graph = new Graph();
+
+        assert.throws(function () {
+          graph.areDirectedNeighbors('source', 'target');
+        }, notFound());
+      },
+
+      'it should correctly return whether two nodes are neighbors.':
+        function () {
+          const graph = new Graph();
+
+          graph.mergeDirectedEdge('Mary', 'Joseph');
+          graph.mergeUndirectedEdge('Martha', 'Mary');
+
+          assert.strictEqual(
+            graph.areDirectedNeighbors('Mary', 'Joseph'),
+            true
+          );
+          assert.strictEqual(
+            graph.areDirectedNeighbors('Joseph', 'Mary'),
+            true
+          );
+          assert.strictEqual(
+            graph.areDirectedNeighbors('Martha', 'Mary'),
+            false
+          );
+
+          const undirectedGraph = new Graph({type: 'undirected'});
+          undirectedGraph.mergeEdge('Mary', 'Martha');
+
+          assert.strictEqual(
+            undirectedGraph.areDirectedNeighbors('Mary', 'Martha'),
+            false
+          );
+        }
+    },
+
+    '#.areInNeighbors': {
+      'it should throw if node is not in the graph.': function () {
+        const graph = new Graph();
+
+        assert.throws(function () {
+          graph.areInNeighbors('source', 'target');
+        }, notFound());
+      },
+
+      'it should correctly return whether two nodes are neighbors.':
+        function () {
+          const graph = new Graph();
+
+          graph.mergeDirectedEdge('Mary', 'Joseph');
+          graph.mergeUndirectedEdge('Martha', 'Mary');
+
+          assert.strictEqual(graph.areInNeighbors('Mary', 'Joseph'), false);
+          assert.strictEqual(graph.areInNeighbors('Joseph', 'Mary'), true);
+          assert.strictEqual(graph.areInNeighbors('Martha', 'Mary'), false);
+
+          const undirectedGraph = new Graph({type: 'undirected'});
+          undirectedGraph.mergeEdge('Mary', 'Martha');
+
+          assert.strictEqual(
+            undirectedGraph.areInNeighbors('Mary', 'Martha'),
+            false
+          );
+        }
+    },
+
+    '#.areOutNeighbors': {
+      'it should throw if node is not in the graph.': function () {
+        const graph = new Graph();
+
+        assert.throws(function () {
+          graph.areOutNeighbors('source', 'target');
+        }, notFound());
+      },
+
+      'it should correctly return whether two nodes are neighbors.':
+        function () {
+          const graph = new Graph();
+
+          graph.mergeDirectedEdge('Mary', 'Joseph');
+          graph.mergeUndirectedEdge('Martha', 'Mary');
+
+          assert.strictEqual(graph.areOutNeighbors('Mary', 'Joseph'), true);
+          assert.strictEqual(graph.areOutNeighbors('Joseph', 'Mary'), false);
+          assert.strictEqual(graph.areOutNeighbors('Martha', 'Mary'), false);
+
+          const undirectedGraph = new Graph({type: 'undirected'});
+          undirectedGraph.mergeEdge('Mary', 'Martha');
+
+          assert.strictEqual(
+            undirectedGraph.areOutNeighbors('Mary', 'Martha'),
+            false
+          );
+        }
+    },
+
+    '#.areOutboundNeighbors': {
+      'it should throw if node is not in the graph.': function () {
+        const graph = new Graph();
+
+        assert.throws(function () {
+          graph.areOutboundNeighbors('source', 'target');
+        }, notFound());
+      },
+
+      'it should correctly return whether two nodes are neighbors.':
+        function () {
+          const graph = new Graph();
+
+          graph.mergeDirectedEdge('Mary', 'Joseph');
+          graph.mergeUndirectedEdge('Martha', 'Mary');
+
+          assert.strictEqual(
+            graph.areOutboundNeighbors('Mary', 'Joseph'),
+            true
+          );
+          assert.strictEqual(
+            graph.areOutboundNeighbors('Joseph', 'Mary'),
+            false
+          );
+          assert.strictEqual(
+            graph.areOutboundNeighbors('Martha', 'Mary'),
+            true
+          );
+
+          const undirectedGraph = new Graph({type: 'undirected'});
+          undirectedGraph.mergeEdge('Mary', 'Martha');
+
+          assert.strictEqual(
+            undirectedGraph.areOutboundNeighbors('Mary', 'Martha'),
+            true
+          );
+        }
+    },
+
+    '#.areInboundNeighbors': {
+      'it should throw if node is not in the graph.': function () {
+        const graph = new Graph();
+
+        assert.throws(function () {
+          graph.areInboundNeighbors('source', 'target');
+        }, notFound());
+      },
+
+      'it should correctly return whether two nodes are neighbors.':
+        function () {
+          const graph = new Graph();
+
+          graph.mergeDirectedEdge('Mary', 'Joseph');
+          graph.mergeUndirectedEdge('Martha', 'Mary');
+
+          assert.strictEqual(
+            graph.areInboundNeighbors('Mary', 'Joseph'),
+            false
+          );
+          assert.strictEqual(graph.areInboundNeighbors('Joseph', 'Mary'), true);
+          assert.strictEqual(graph.areInboundNeighbors('Martha', 'Mary'), true);
+
+          const undirectedGraph = new Graph({type: 'undirected'});
+          undirectedGraph.mergeEdge('Mary', 'Martha');
+
+          assert.strictEqual(
+            undirectedGraph.areInboundNeighbors('Mary', 'Martha'),
+            true
+          );
+        }
+    },
+
+    '#.areUndirectedNeighbors': {
+      'it should throw if node is not in the graph.': function () {
+        const graph = new Graph();
+
+        assert.throws(function () {
+          graph.areUndirectedNeighbors('source', 'target');
+        }, notFound());
+      },
+
+      'it should correctly return whether two nodes are neighbors.':
+        function () {
+          const graph = new Graph();
+
+          graph.mergeDirectedEdge('Mary', 'Joseph');
+          graph.mergeUndirectedEdge('Martha', 'Mary');
+
+          assert.strictEqual(
+            graph.areUndirectedNeighbors('Mary', 'Joseph'),
+            false
+          );
+          assert.strictEqual(
+            graph.areUndirectedNeighbors('Joseph', 'Mary'),
+            false
+          );
+          assert.strictEqual(
+            graph.areUndirectedNeighbors('Martha', 'Mary'),
+            true
+          );
+
+          const directedGraph = new Graph({type: 'directed'});
+          directedGraph.mergeEdge('Mary', 'Martha');
+
+          assert.strictEqual(
+            directedGraph.areUndirectedNeighbors('Mary', 'Martha'),
+            false
+          );
+        }
+    },
+
+    '#.areNeighbors': {
+      'it should throw if node is not in the graph.': function () {
+        const graph = new Graph();
+
+        assert.throws(function () {
+          graph.areNeighbors('source', 'target');
+        }, notFound());
+      },
+
+      'it should correctly return whether two nodes are neighbors.':
+        function () {
+          const graph = new Graph();
+
+          graph.mergeDirectedEdge('Mary', 'Joseph');
+          graph.mergeUndirectedEdge('Martha', 'Mary');
+
+          assert.strictEqual(graph.areNeighbors('Mary', 'Joseph'), true);
+          assert.strictEqual(graph.areNeighbors('Joseph', 'Mary'), true);
+          assert.strictEqual(graph.areNeighbors('Martha', 'Mary'), true);
+          assert.strictEqual(graph.areNeighbors('Joseph', 'Martha'), false);
+
+          const undirectedGraph = new Graph({type: 'undirected'});
+          undirectedGraph.mergeEdge('Mary', 'Martha');
+
+          assert.strictEqual(
+            undirectedGraph.areNeighbors('Mary', 'Martha'),
+            true
+          );
+        }
+    },
+
     '#.source': {
       'it should throw if the edge is not in the graph.': function () {
         const graph = new Graph();